requirements.txt
flask==3.0.3
dnslib==0.9.23
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.0
requests==2.31.0
google-generativeai==0.8.2
openai==1.47.0
datasets==3.0.0
python-dotenv==1.0.1
Flask-CORS==4.0.0
